requirements.txt
gradio==3.41.2
modules==1.0.0
Pillow==9.5.0
pydantic==1.10.12
python_box==7.1.1
rembg==2.0.52
Requests==2.31.0
uvicorn==0.23.2
fastapi==0.94.0